mv accumulator out
