+++ datafusion/functions-aggregate-common/src/lib.rs
@@ -15,6 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
+pub mod accumulator;
 pub mod aggregate;
 pub mod merge_arrays;
 pub mod order;
